+++ src/components/Card/detailedCard/detailedCard.module.scss
@@ -41,6 +41,7 @@
 }
 
 .image {
+  margin: 0 auto;
   max-width: 300px;
   height: 200px;
 }
