+++ src/utils/constants/Constants.ts
@@ -1,1 +1,1 @@
-export const SEARCH_LOCALSTORAGE_KEY = 'react-rss-vikkitorry-version.2.2';
+export const SEARCH_LOCALSTORAGE_KEY = 'react-rss-vikkitorry-version.2.22';
